4/scripts/requirements.txt
click==8.1.7
colorama==0.4.6
markdown-it-py==3.0.0
mdurl==0.1.2
paho-mqtt==2.0.0
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.0
shellingham==1.5.4
typer==0.9.0
typing_extensions==4.9.0
